Fixed bug in GUI involving changing a name.
When changing a name of a task in the GUI, it would be impossible to remove it later - although the task would get removed from the list of tasks. This caused strange situations.

Apparently, when changing an alement in a set, you have to reinsert the element. Otherwise, when you iterate over the elments in the set later, you will find that the following construct gives unexpected results:

    // Note: createdTasks is a HashSet<Task>
    for (Task t : createdTasks) {
      System.out.println(createdTasks.contains(t)); // may print false!
    }

This is of course very strange behaviour, but not a bug. To quote the JavaDoc of Set:

> Note: Great care must be exercised if mutable objects are used as set elements.
> The behavior of a set is not specified if the value of an object is changed in
> a manner that affects equals comparisons while the object is an element in the
> set. A special case of this prohibition is that it is not permissible for a set
> to contain itself as an element.

Learned something new today, jaj.